rename : test 코드 디렉토리 구조에 맞게 파일 구조 변경
