+++ habit/src/test/java/app/habit/service/HabitAssessmentManagerServiceTest.java
@@ -1,4 +1,4 @@
-package app.habit.service.gpt;
+package app.habit.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
